+++ src/main/scala/de/htwg/winesmeeper/Controller/Commands/FlagCommand.scala
@@ -14,17 +14,23 @@ case class FlagCommand(ctrl: Controller, x: Int, y: Int) extends Command:
     if (!f.isOpened) then
       ctrl.gb = Board(ctrl.gb.board.updated(x, ctrl.gb.board(x).updated(y, Field(f.isBomb, f.isOpened, !f.isFlag))))
       true
-    else
-      print("hi2");false
+    else false
 
   override def undoStep(): Boolean = doStep()
 
   override def redoStep(): Boolean = doStep()
+
+  override def toString: String = f"flag($x, $y)"
   
-object CORFlag extends CommandCOR:
+object FlagCOR extends CommandCOR:
   override val cmd = "flag"
-  override val helpMsg: String = "sets a flag at the given coordinate"
+  override val helpMsg: String = "flag or unflag the given coordinate"
   override val next: CommandCOR = OpenFieldCOR
+  override val specHelpMsg: String =
+    """flag <x> <y>:
+      |  mark this position as flag or remove the flag
+      |""".stripMargin
 
   override def buildCmd(cmd: String, x: Int, y: Int, ctrl: Controller): Try[Command] =
     if cmd == this.cmd then Success(FlagCommand(ctrl, x, y)) else next.buildCmd(cmd,x,y,ctrl)
+