+++ src/main/scala/de/htwg/winesmeeper/Controller/Commands/OpenFieldCmd.scala
@@ -7,6 +7,7 @@ import scala.util.{Success, Try}
 
 case class OpenFieldCmd(ctrl: Controller, x: Int, y: Int) extends Command:
 
+  val isFlag: Boolean = ctrl.gb.board(x)(y).isFlag
   override def doStep(): Boolean =
     step(true)
 
@@ -21,7 +22,7 @@ case class OpenFieldCmd(ctrl: Controller, x: Int, y: Int) extends Command:
     val f = gb.getFieldAt(x, y)
     if discover == f.isOpened then false
     else
-      val newVector = gb.board.updated(x, gb.board(x).updated(y, Field(f.isBomb, discover)))
+      val newVector = gb.board.updated(x, gb.board(x).updated(y, Field(f.isBomb, discover, !discover && isFlag)))
       ctrl.gb = new Board(newVector)
       if f.isBomb && !discover then ctrl.changeState("running")
       if f.isBomb && discover then ctrl.changeState("lose");
@@ -34,10 +35,18 @@ case class OpenFieldCmd(ctrl: Controller, x: Int, y: Int) extends Command:
       if ctrl.isVictory then ctrl.changeState("win")
       true
 
+  override def toString: String = f"open($x, $y)"
+
 object OpenFieldCOR extends CommandCOR:
   override val cmd = "open"
   override val helpMsg = "opens the field of the given coordinate"
   override val next: CommandCOR = LastElemCmdCOR
+  override val specHelpMsg: String =
+    """open <x> <y>:
+      |  Opens a field and if you hit a bomb, you loose!
+      |  But no pressure you can undo your fault!
+      |""".stripMargin
 
   override def buildCmd(cmd: String, x: Int, y: Int, ctrl: Controller): Try[Command] =
     if cmd == this.cmd then Success(OpenFieldCmd(ctrl, x, y)) else next.buildCmd(cmd,x,y,ctrl)
+