+++ src/main/scala/de/htwg/winesmeeper/Controller/SysCommands/HelpCmd.scala
@@ -2,17 +2,33 @@ package de.htwg.winesmeeper.Controller.SysCommands
 
 import de.htwg.winesmeeper.Controller.Controller
 import de.htwg.winesmeeper.Controller.Commands
-import de.htwg.winesmeeper.Controller.Commands.CORFlag
+import de.htwg.winesmeeper.Controller.Commands.FlagCOR
 
 object HelpCmd extends SysCommandCOR:
   override val cmd: String = "help"
   override val helpMsg: String = "made this message"
-  override val next: SysCommandCOR = QuitCmd
+  override val next: SysCommandCOR = LoadCmd
+  override val specHelpMsg: String =
+    """help:
+      |  prints an overview of all commands
+      |
+      |help <cmd>:
+      |  prints a specific help message for this message
+      |""".stripMargin
 
-  override def execute(ctrl: Controller): String =
+
+  override def execute(ctrl: Controller, cmd: String, params: Vector[String]): String =
+    val command = if params.length > 1 then SysCommandManager.getAbstractCmd(params(1)) else None
+    command match
+      case Some(value) => value.specHelpMsg
+      case None => standardHelp
+
+  private def standardHelp: String =
    val sysCmds = HelpCmd.listCmds
    val scString = (for cmd <- sysCmds yield s"  ${cmd.cmd}: ${cmd.helpMsg}").mkString("\n")
-   val cmds = CORFlag.listCmds
+   val cmds = FlagCOR.listCmds
    val cmdString = (for cmd <- cmds yield s"  ${cmd.cmd}: ${cmd.helpMsg}").mkString("\n")
-   s"Commands without parameters:\n$scString \n\nCommands with coordinates:\n$cmdString"
+   s"Commands without parameters:\n$scString \n\nCommands with coordinates:\n$cmdString\n\nFor more help type: help + wanted command"
+
+