+++ src/test/scala/main/de/htwg/winesmeeper/tests/Controller/ControllerSpec.scala
@@ -27,7 +27,7 @@ class ControllerSpec extends AnyWordSpec with Matchers:
       ctrl.gameState shouldBe "running"
       val w = Controller(10, 10, 5, 5, 91)
       w.gameState shouldBe "win"
-      val l = Controller(10, 10, 5, 5, 90); l.turn("open",1524, 1243); l.turn("open", 1, 1)
+      val l = Controller(10, 10, 5, 5, 90); l.turn("open", 1524, 1243); l.turn("open", 1, 1)
       l.gameState shouldBe "lose"
       l.getBoard